moved rbyz auxiliary logic to separate file
